# Add optimized ReadAll function