--- a/api/server_windows_test.go
+++ b/api/server_windows_test.go
@@ -52,7 +52,7 @@ func TestNamedPipe(t *testing.T) {
 		},
 	}
 
-	// nolint:noctx // for testing purposes
+	//nolint:noctx // for testing purposes
 	r, err := c.Get("http://npipe/echo-hello")
 	require.NoError(t, err)
 	defer r.Body.Close()