--- a/file/helper_test.go
+++ b/file/helper_test.go
@@ -20,7 +20,6 @@
 package file
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -29,19 +28,19 @@ import (
 )
 
 func TestSafeFileRotateExistingFile(t *testing.T) {
-	tempdir, err := ioutil.TempDir("", "")
+	tempdir, err := os.MkdirTemp("", "")
 	assert.NoError(t, err)
 	defer func() {
 		assert.NoError(t, os.RemoveAll(tempdir))
 	}()
 
 	// create an existing registry file
-	err = ioutil.WriteFile(filepath.Join(tempdir, "registry"),
+	err = os.WriteFile(filepath.Join(tempdir, "registry"),
 		[]byte("existing filebeat"), 0x777)
 	assert.NoError(t, err)
 
 	// create a new registry.new file
-	err = ioutil.WriteFile(filepath.Join(tempdir, "registry.new"),
+	err = os.WriteFile(filepath.Join(tempdir, "registry.new"),
 		[]byte("new filebeat"), 0x777)
 	assert.NoError(t, err)
 
@@ -50,13 +49,13 @@ func TestSafeFileRotateExistingFile(t *testing.T) {
 		filepath.Join(tempdir, "registry.new"))
 	assert.NoError(t, err)
 
-	contents, err := ioutil.ReadFile(filepath.Join(tempdir, "registry"))
+	contents, err := os.ReadFile(filepath.Join(tempdir, "registry"))
 	assert.NoError(t, err)
 	assert.Equal(t, []byte("new filebeat"), contents)
 
 	// do it again to make sure we deal with deleting the old file
 
-	err = ioutil.WriteFile(filepath.Join(tempdir, "registry.new"),
+	err = os.WriteFile(filepath.Join(tempdir, "registry.new"),
 		[]byte("new filebeat 1"), 0x777)
 	assert.NoError(t, err)
 
@@ -64,13 +63,13 @@ func TestSafeFileRotateExistingFile(t *testing.T) {
 		filepath.Join(tempdir, "registry.new"))
 	assert.NoError(t, err)
 
-	contents, err = ioutil.ReadFile(filepath.Join(tempdir, "registry"))
+	contents, err = os.ReadFile(filepath.Join(tempdir, "registry"))
 	assert.NoError(t, err)
 	assert.Equal(t, []byte("new filebeat 1"), contents)
 
 	// and again for good measure
 
-	err = ioutil.WriteFile(filepath.Join(tempdir, "registry.new"),
+	err = os.WriteFile(filepath.Join(tempdir, "registry.new"),
 		[]byte("new filebeat 2"), 0x777)
 	assert.NoError(t, err)
 
@@ -78,7 +77,7 @@ func TestSafeFileRotateExistingFile(t *testing.T) {
 		filepath.Join(tempdir, "registry.new"))
 	assert.NoError(t, err)
 
-	contents, err = ioutil.ReadFile(filepath.Join(tempdir, "registry"))
+	contents, err = os.ReadFile(filepath.Join(tempdir, "registry"))
 	assert.NoError(t, err)
 	assert.Equal(t, []byte("new filebeat 2"), contents)
 }