--- a/loader/config.go
+++ b/loader/config.go
@@ -20,7 +20,6 @@ package loader
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 
 	"gopkg.in/yaml.v2"
@@ -100,7 +99,7 @@ func NewConfigFrom(from interface{}, opts ...interface{}) (*Config, error) {
 		if closer, ok := from.(io.Closer); ok {
 			defer closer.Close()
 		}
-		fData, err := ioutil.ReadAll(in)
+		fData, err := io.ReadAll(in)
 		if err != nil {
 			return nil, err
 		}