--- a/loader/config_test.go
+++ b/loader/config_test.go
@@ -18,7 +18,6 @@
 package loader
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -56,7 +55,7 @@ func TestInputsResolveNOOP(t *testing.T) {
 		},
 	}
 
-	tmp, err := ioutil.TempDir("", "config")
+	tmp, err := os.MkdirTemp("", "config")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmp)
 
@@ -87,7 +86,7 @@ func testToMapStr(t *testing.T) {
 }
 
 func testLoadFiles(t *testing.T) {
-	tmp, err := ioutil.TempDir("", "watch")
+	tmp, err := os.MkdirTemp("", "watch")
 	require.NoError(t, err)
 	defer os.RemoveAll(tmp)
 
@@ -132,6 +131,6 @@ func testLoadFiles(t *testing.T) {
 func dumpToYAML(t *testing.T, out string, in interface{}) {
 	b, err := yaml.Marshal(in)
 	require.NoError(t, err)
-	err = ioutil.WriteFile(out, b, 0600)
+	err = os.WriteFile(out, b, 0600)
 	require.NoError(t, err)
 }