--- a/loader/discover_test.go
+++ b/loader/discover_test.go
@@ -18,7 +18,6 @@
 package loader
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -72,7 +71,7 @@ func TestDiscover(t *testing.T) {
 
 func withFiles(files []string, fn func(dst string, t *testing.T)) func(t *testing.T) {
 	return func(t *testing.T) {
-		tmp, _ := ioutil.TempDir("", "watch")
+		tmp, _ := os.MkdirTemp("", "watch")
 		defer os.RemoveAll(tmp)
 
 		for _, file := range files {